strands_from_scratch/03-resturant-agent-deployto-AgentCore/requirements.txt
boto3
botocore
awscli
opensearch-py
requests-aws4auth
pyyaml
retrying
pandas
strands-agents
strands-agents-tools
uv
bedrock-agentcore
bedrock-agentcore-starter-toolkit